api/requirements.txt
flask==3.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.8
